refactor: pulling all v1 auth into the v1 file
