+++ packages/shared/document-model/auth.ts
@@ -1,15 +1,14 @@
 import { z } from "zod";
 import { createAction, type Action } from "./actions.js";
 import {
+  assertValidGrantUpsert,
+  assertValidInitialGrants,
+  evaluateGrants,
   GrantSchema,
-  grantProblem,
-  GroupPrincipalNotAllowedError,
-  InvalidGrantError,
   isPlainValue,
   MAX_AUTH_GRANTS,
 } from "./auth-v1.js";
 import { base58Decode, base64UrlToBytes } from "./crypto.js";
-import { groupDocumentType } from "./document-type.js";
 import type { PHDocument } from "./documents.js";
 import {
   AuthActionNotAllowedError,
@@ -22,11 +21,9 @@ import {
 } from "./errors.js";
 import {
   createAuthState,
-  type Capability,
   type Grant,
   type PHAuthState,
   type PHBaseState,
-  type Principal,
 } from "./state.js";
 
 // --- Action types --------------------------------------------------------
@@ -166,22 +163,6 @@ function assertActionInputShape(input: unknown): void {
   }
 }
 
-/** V1 shape rules plus the group-document group-principal ban. */
-function assertValidGrant(grant: unknown, documentType: string): void {
-  const grantId =
-    isPlainValue(grant) && typeof grant.id === "string" ? grant.id : "";
-  const problem = grantProblem(grant);
-  if (problem !== null) {
-    throw new InvalidGrantError(grantId, problem);
-  }
-  if (
-    documentType === groupDocumentType &&
-    "group" in (grant as Grant).principal
-  ) {
-    throw new GroupPrincipalNotAllowedError(grantId);
-  }
-}
-
 function withGrants<TState extends PHBaseState>(
   document: PHDocument<TState>,
   grants: Grant[],
@@ -275,12 +256,7 @@ export function applyInitializeAuthAction<TState extends PHBaseState>(
   if (!Array.isArray(grants)) {
     throw new InvalidActionInputError({ grants: "must be an array" });
   }
-  if (grants.length > MAX_AUTH_GRANTS) {
-    throw new InvalidGrantError("", `policy exceeds ${MAX_AUTH_GRANTS} grants`);
-  }
-  for (const grant of grants) {
-    assertValidGrant(grant, document.header.documentType);
-  }
+  assertValidInitialGrants(grants, document.header.documentType);
   const creatorKey = document.header.sig.publicKey;
   const signerKey = action.context?.signer?.app.key;
   // Any key material marks a signed header. Unsupported key types then fail
@@ -308,15 +284,9 @@ export function applySetGrantAction<TState extends PHBaseState>(
 ): PHDocument<TState> {
   assertActionInputShape(action.input);
   const { grant } = action.input;
-  assertValidGrant(grant, document.header.documentType);
   const grants = document.state.auth.grants;
+  assertValidGrantUpsert(grant, grants, document.header.documentType);
   const exists = grants.some((g) => g.id === grant.id);
-  if (!exists && grants.length >= MAX_AUTH_GRANTS) {
-    throw new InvalidGrantError(
-      grant.id,
-      `policy exceeds ${MAX_AUTH_GRANTS} grants`,
-    );
-  }
   const next = exists
     ? grants.map((g) => (g.id === grant.id ? grant : g))
     : [...grants, grant];
@@ -441,53 +411,11 @@ export type AuthSubject = {
 
 export type AuthDecision = "allow" | "deny";
 
-function capabilityCovers(
-  capability: Capability,
-  request: AuthRequest,
-): boolean {
-  if (capability.can !== request.verb) {
-    return false;
-  }
-  const scope = capability.scope;
-  if (scope !== undefined && scope !== "*" && scope !== request.scope) {
-    return false;
-  }
-  if (capability.can === "execute") {
-    // An execute capability with no operation list covers every operation in the scope.
-    if (capability.operation === undefined) {
-      return true;
-    }
-    return (
-      request.operation !== undefined &&
-      capability.operation.includes(request.operation)
-    );
-  }
-  return true;
-}
-
-function principalMatches(principal: Principal, subject: AuthSubject): boolean {
-  if ("anyone" in principal) {
-    return true;
-  }
-  if ("address" in principal) {
-    return (
-      subject.address !== undefined &&
-      subject.address.toLowerCase() === principal.address.toLowerCase()
-    );
-  }
-  // { group } and { match } are not evaluated yet: group membership needs the
-  // PHGroup model (a missing group never widens access) and conditions are deferred.
-  return false;
-}
-
 /**
  * Evaluates the auth policy for a single request. Pure and deterministic.
  *
  * An uninitialized policy (version 0, or absent auth state) leaves the document
  * open. Once a policy exists the default is deny, and grants stack in order.
- *
- * Group and match principals and `where` conditions are not evaluated yet; a
- * grant that uses any of them never applies.
  */
 export function decide(
   auth: PHAuthState | undefined,
@@ -512,18 +440,5 @@ export function decide(
     return "deny";
   }
 
-  let decision: AuthDecision = "deny";
-  for (const grant of auth.grants) {
-    // `where` is not evaluated yet; a conditional grant never applies.
-    if (grant.where !== undefined) {
-      continue;
-    }
-    if (
-      capabilityCovers(grant.capability, request) &&
-      principalMatches(grant.principal, subject)
-    ) {
-      decision = grant.effect;
-    }
-  }
-  return decision;
+  return evaluateGrants(auth.grants, subject, request);
 }
